Rename to .modelfarm
