+++ src/replit/tests/ai/modelfarm/google/language_models/test_text_generation_model.py
@@ -1,5 +1,5 @@
 import pytest
-from replit.ai.google.language_models import TextGenerationModel, TextGenerationResponse
+from replit.ai.modelfarm.google.language_models import TextGenerationModel, TextGenerationResponse
 
 TEST_PARAMETERS = {
     "temperature": 0.5,  # Temperature controls the degree of randomness in token selection.
