+++ src/replit/tests/ai/modelfarm/test_config.py
@@ -1,4 +1,4 @@
-from replit.ai.config import initialize, get_config, Config
+from replit.ai.modelfarm.config import initialize, get_config, Config
 
 
 def test_config_initialization():
