+++ src/replit/tests/ai/modelfarm/test_embed_model.py
@@ -1,6 +1,6 @@
 import pytest
 from replit.ai import EmbedModel
-from replit.ai.exceptions import BadRequestException
+from replit.ai.modelfarm.exceptions import BadRequestException
 
 # module level constants
 CONTENT = [{"content": "1 + 1 = "}]
