+++ src/replit/tests/ai/modelfarm/test_identity.py
@@ -7,9 +7,9 @@ from unittest.mock import patch
 from unittest.mock import Mock
 import pyseto
 import pytest
-from replit.ai.identity import verify
-from replit.ai.identity.sign import SigningAuthority
-from replit.ai.replit_identity_token_manager import ReplitIdentityTokenManager
+from replit.ai.modelfarm.identity import verify
+from replit.ai.modelfarm.identity.sign import SigningAuthority
+from replit.ai.modelfarm.replit_identity_token_manager import ReplitIdentityTokenManager
 
 from replit.ai import config
